MTA-743 move payment type URI to its own class
