+++ routes/teacher.php
@@ -24,7 +24,7 @@ Route::group(['middleware' => ['auth', 'teacher', 'subscribed']], function () {
         Route::resource('holiday', 'HolidayController');
         Route::post('student-save', 'StudentController@store');
         Route::get('status', 'ReportController@status');
-        Route::get('payment-types', 'InvoiceController@getPaymentTypes');
+        Route::get('payment-types', 'PaymentTypeController@index');
         Route::resource('teacher-settings', 'TeacherSettingsController');
         Route::get('student/lesson/{id}/{day?}', 'StudentLessonController@getStudent');
         Route::post('student/lessons', 'StudentLessonController@store');
