+++ tests/Feature/Http/Controllers/InvoiceControllerTest.php
@@ -7,7 +7,6 @@ use App\Models\PaymentType;
 use App\Models\Student;
 use App\Models\User;
 use Illuminate\Foundation\Testing\RefreshDatabase;
-use PaymentTypeTableSeeder;
 use Tests\TestCase;
 
 class InvoiceControllerTest extends TestCase
@@ -17,64 +16,6 @@ class InvoiceControllerTest extends TestCase
     public function setUp(): void
     {
         parent::setUp();
-    }
-
-    public function test_payment_types_json()
-    {
-        $this->withoutMiddleware();
-
-        $this->seed(PaymentTypeTableSeeder::class);
-
-        $response = $this->get('/web/payment-types');
-
-        $response->assertOk();
-        $response->assertJsonCount(11);
-        $response->assertJson([
-            [
-                'id' => 1,
-                'name' => "Cash",
-            ],
-            [
-                'id' => 2,
-                'name' => "Check",
-            ],
-            [
-                'id' => 3,
-                'name' => "Credit Card",
-            ],
-            [
-                'id' => 4,
-                'name' => "Cash App",
-            ],
-            [
-                'id' => 5,
-                'name' => "Apple Pay",
-            ],
-            [
-                'id' => 6,
-                'name' => "Google Pay",
-            ],
-            [
-                'id' => 7,
-                'name' => "Stripe",
-            ],
-            [
-                'id' => 8,
-                'name' => "PayPal",
-            ],
-            [
-                'id' => 9,
-                'name' => "Venmo",
-            ],
-            [
-                'id' => 10,
-                'name' => "Zelle",
-            ],
-            [
-                'id' => 11,
-                'name' => "Other",
-            ],
-        ]);
     }
 
     public function test_index_page_route_view()
